Implement the SecretReference type.
References are no longer passed around as strings and validated as needed.
They are now strictly parsed and validity is enforced by the type.
The template renderer will now parse secret references directly, which has the side effect that the template renderer does now perform some small allocations where it previously was no-allocation.
The SecretsProvider trait now has a ReferenceParser subtrait which must be implemented for every provider.
This does impact testability in some ways, as making a mock parser is not possible anymore.